Creating Window object & Popover rework First draft of the new BrowserWindow object. So the users can be get easilier to the window events and connect it with the fritz2 handler.
A new WindowListener and his functions created in listener.kt so the events of the Window can be connect to the listener.

Refactoring Listener to DomListener
Creating interface for Listener Methods
Creating test for window.kt

The Popover is the first component which use the new BrowserWindow object.
It has two new options (closeOnBlur, closeOnEscape) these options works with the browser window events.