Add debug information for headless components
In order to improve the usage of headless components all components and its bricks will render out HTML comments that name their corresponding component or brick name. This way the matching between the Kotlin names and its HTML equivalents is much easier.

Most hint comments are located as direct predecessor of the created HTML element. For all bricks that are based upon some `Flow` this is not possible due to their managed nature. In those cases the comment appears as first sub-node within the created element and its text startes with "parent is xyz" to clarify its relationship.

In order to activate those helpful "debugging" information, one must put the `HEADLESS_DEBUG` key into the scope with a `true` value.

fixes #626