Add default data-binding as fallback for headless components
As for some components it is absolutely mandatory to have some data-binding, in order to avoid some NPE we have added default appropriate data-bindings for those. Additionally, some warning messages will be logged into the console in such a fallback case, so the user gets a hint, what he has forgotten to set.