Repair aria-haspopup for PopUpPanel based components
Former to this PR components based upon PopoverPanel set the ``aria-haspopup`` property always to true.

According to its [definition](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-haspopup#description) the value only might be ``true`` if the component represents some "menu". In other cases there are dedicated values, that should be set:
- menu
- listbox
- tree
- grid
- dialog

This PR adds a parameter for this value to the ``PopUpPanel``, so that every component has to pass a value.

The following headless components now sets the following values accordingly:
- ``listbox`` -> ``listbox``
- ``menu`` -> ``menu``
- ``popOver`` -> ``dialog``
- ``tooltip`` -> ``dialog``

The attribute is set defensively, so that a user can always overrule those defaults by applying the fitting one into the ``reference`` tag of the panel.