Add default z-Index for PopUpPanel
I order to render the floating content of a PopUpPanel based component, we have defined a static z-Index with value ``30`` for the popper div, which can't be styled in a safe and reliable way from the outside.

This value corresponds to the median value of the predefined [tailwindcss](https://tailwindcss.com/docs/z-index) z-indices.