Improve focus-trap
Add some reactive variant named ``trapFocusWhenever`` to the existing ``trapFocus`` function, which requires some boolean flow as first parameter that acts as a conditional decider: Only on ``true`` values on the flow the focus-trap will get established; on ``false`` values it will get disabled.

This is very useful for situations, where some kind of "open" state (as for ``PopUpPanel`` based components for example) triggers the visibility of a content, where the trap should be active. After closing however, the tab should be freed again, so that the user can continue to tab through the application.

Also removes unnecessary tab index from tab-panel.