Fix renderEach behaviour after regression by RC3 release
The RC3 Release improved the rendering by dropping unnecessary mount-operations by PR #735.

Because of the changes, a regression appeared, which was caused by the shape of the `Patch`-implementations for the `renderEach`-functions: They were *data* classes and therefor it is quite easy to create equal patches, which then would be dropped by the mount-optimization! Imagine deleting the first item of a list at least twice for example.

This PR fixes the regression, so that a patch is always not equal to any other one (dropping data classes in favour of regular classes with object identity for equals).

In order to prevent a future regression some dedicated unit tests were added too.