Substitute tailwindcss class with vanilla CSS
There was a small remnant from the headless development within the popover component factory, that uses the `relative` class from tailwindcss - this is obviously wrong, as fritz2 is totally agnostic of CSS frameworks. So it is replaced now by some custom fritz2 prefixed class, that simply sets the `position=relative` too.