Fixes an issue with overlapping validation messages of mapped stores
The filtering of mapped validation messages used to test only the prefix. This could lead to false assigned messages, if a field name of some lens is exactly the prefix of another one. So messages for the field `foo` would also appear in a mapped store of field `foobar` as both share the same prefix.

This is now fixed, as the prefix check now excludes leave nodes.