Removes portal stylings
The default `z-index`-Stylings are removed entirely from the portal-factory. As the portal-containers can be defined by the user, there is no more need to "hide" them with `display-contents` and implement some `z-index` behavior for their children.

This is a huge step to consequently remain headless and be agnostic of any styling!