Improves the validation message filtering
Up to the `RC5 / RC6` version of fritz2, the `messages()`-function on `Store`s just matched the *exact* path. This behaviour was changed by those release candidates: Right now all messages that matches the exact path, but also all messages machining the sub-path will be collected.

Our experience has shown, that this behaviour change was a mistake, as the 80% use case is to collect all messages for an exact path. Most of the time one has dedicated form elements, that fit exactly one property of a model and thus should only show messages for itself. It is very rarely happening, that all messages of a sub-model should be shown all together.

This is why this PR changes again the default behaviour, but also introduces a dedicated variant for collecting all sub-tree messages with the new `messagesOfSubTree()`-function.

Also, we expose the `messages(filterPredicate: (M) -> Boolean)`-function, which both other functions uses internally. You can use this function apply custom filtering to the messages.