Fix Selection Behaviour of ListBox
Due to PR #833 the closing mechanism of headless listbox has changed, so that clicking / pressing enter or space on the previously selected item did not close the listbox anymore. This commit restores the previous and desired behaviour again.

In order to avoid future regressions, a dedicated playwright test is added.