Make lenses genration more resilient with KSP >= 1.9.x
Obviously the KSP based LensesAnnotationProcessor was more resilient in the past before we upgraded to Kotlin 1.9.x and appropriate KSP with RC15.

We encountered some issues with generating lenses that only came with this bad warning:
```
w: [ksp] Unable to process:dev.fritz2.lens.LensesProcessor:   SomeClass
```

This was due to some not resolvable types inside the annotated data class, like calls to functions from other libs or delegated properties. This PR fixes such issues, as we ignore everything from the annotated class but the primary constructor. The latter is the only interesting piece for our lens generation.