Reduces flickering when re-opening a combobox
The flickering occurred, when a combobox was re-opened after a previous
query was performed and an item has been selected.

This is fixed by allowing the dropdown's content to re-render when closed.
The number of calculations stays roughly the same as the results are only
computed when the input data changed.