Combobox: Only accept distinct values via data-binding (#913)
This commit adjust the combobox's data-binding to only accept values
that are different from the previous selection.

In most cases, this should not make a difference.
There are corner cases, however, where the value is quickly updated
multiple times in a row, causing the internals to run into
race-conditions. The above changes should prevent this issue.