Improves support for upcasting lenses
Introduces a dedicated `Lens`-factory fpr upcasting lenses:
`lensForUpcasting`

Its inner code will catch `ClassCastException`s, which can
occur in nested `Mountpoints`, and throw a dedicated `CollectionLensGetException`.
This solution is similar to the handling of false index access or
access by keys for collections, we already have integrated.

This is necessary due to the fact, that fritz2 does not maintain
any sort of `Flow`-hierarchie from the outermost to the innermost
renderings. So the framework cannot control the sequence of collecting
the involved flows, in order to prevent outdated data or - for this case - *types*.

This is a practical compromise, that will work very well for UIs,
as the visual appearance is kind of *dictated* by the properties of the underlying type.