Improves messages of `CollectionLensGetException`
This PR adds a `message`-property to the `CollectionLensGetException`-type. This message property is then used inside the various `lens*`-factories for collections in order to provide some useful hints. For example the `index` or `id` is composed alongside with the name of the factory.

This is a huge improvement to trace the error!